Move some type declarations to a new file
