+++ src/lib/features/store.ts
@@ -1,153 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-/**
- * Feature lifecycle stages
- */
-export enum FeatureStage {
-  /**
-   * Alpha features are early/experimental
-   * - Disabled by default
-   * - May contain bugs or be incomplete
-   * - May change significantly or be removed
-   * - Not recommended for production
-   */
-  ALPHA = "alpha",
-
-  /**
-   * Beta features are maturing but not fully stable
-   * - Enabled by default
-   * - Better tested but may still have issues
-   * - API may still undergo minor changes
-   * - Use with caution in production
-   */
-  BETA = "beta",
-
-  /**
-   * GA (General Availability) features are stable
-   * - Always enabled (when available in current version)
-   * - API is stable and maintained
-   * - Suitable for production use
-   */
-  GA = "ga",
-}
-
-/**
- * Metadata for a feature flag
- */
-export interface FeatureMetadata {
-  /** Human-readable name of the feature */
-  name: string;
-
-  /** Description of what the feature does */
-  description: string;
-
-  /** The feature's lifecycle stage */
-  stage: FeatureStage;
-
-  /** Semver version when the feature was introduced */
-  since: string;
-
-  /**
-   * Semver version when the feature will be removed
-   * null means no planned removal
-   */
-  until: string | null;
-
-  /** Default value based on stage if not explicitly set */
-  defaultValue: boolean;
-}
-
-/**
- * Information about a feature including its key and metadata
- */
-export interface FeatureInfo {
-  key: string;
-  metadata: FeatureMetadata;
-}
-
-/**
- * Feature documentation structure
- */
-export interface FeaturesDoc {
-  currentVersion: string;
-  features: Record<
-    string,
-    {
-      name: string;
-      description: string;
-      stage: string;
-      since: string;
-      until: string | null;
-      isAvailable: boolean;
-      isEnabled: boolean;
-      value: FeatureValue | undefined;
-    }
-  >;
-}
-
-type FeatureValue = string | boolean | number;
-
-/**
- * All known feature flags with their metadata
- */
-export const FeatureFlags: Record<string, FeatureInfo> = {
-  DEBUG_MODE: {
-    key: "debug_mode",
-    metadata: {
-      name: "Debug Mode",
-      description: "Enables verbose logging and debugging features",
-      stage: FeatureStage.GA,
-      since: "1.0.0",
-      until: null,
-      defaultValue: true,
-    },
-  },
-  EXPERIMENTAL_API: {
-    key: "experimental_api",
-    metadata: {
-      name: "Experimental API",
-      description: "Enables experimental APIs that may change",
-      stage: FeatureStage.ALPHA,
-      since: "1.2.0",
-      until: null,
-      defaultValue: false,
-    },
-  },
-  PERFORMANCE_METRICS: {
-    key: "performance_metrics",
-    metadata: {
-      name: "Performance Metrics",
-      description: "Enables collection and reporting of performance metrics",
-      stage: FeatureStage.BETA,
-      since: "1.1.0",
-      until: null,
-      defaultValue: true,
-    },
-  },
-  BETA_FEATURES: {
-    key: "beta_features",
-    metadata: {
-      name: "Beta Features Bundle",
-      description: "Enables all current beta features as a bundle",
-      stage: FeatureStage.BETA,
-      since: "1.0.0",
-      until: null,
-      defaultValue: true,
-    },
-  },
-  CHARLIE_FEATURES: {
-    key: "charlie_features",
-    metadata: {
-      name: "Charlie Features",
-      description: "Legacy feature bundle being phased out",
-      stage: FeatureStage.GA,
-      since: "1.3.0",
-      until: "1.9.0",
-      defaultValue: true,
-    },
-  },
-};
+import { FeatureFlags, FeatureValue } from "./FeatureFlags";
+import { FeatureInfo, FeatureMetadata, FeatureStage, FeaturesDoc } from "./FeatureTypes";
 
 /**
  * Global store for feature flags
